backend/requirements.txt
crewai==0.11.0
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.1
langchain==0.1.0
langchain-openai==0.0.2
openai==1.12.0
python-engineio==4.8.0
python-socketio==5.11.1
eventlet==0.35.2